+++ team/src/main/java/sansam/team/project/command/application/dto/project/ProjectMemberUpdateDTO.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import sansam.team.common.YnType;
+import sansam.team.common.aggregate.YnType;
 
 @Getter
 @Setter
