+++ team/src/main/java/sansam/team/project/command/domain/aggregate/entity/ProjectMember.java
@@ -3,8 +3,8 @@ package sansam.team.project.command.domain.aggregate.entity;
 import jakarta.persistence.*;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import sansam.team.common.BaseTimeEntity;
-import sansam.team.common.YnType;
+import sansam.team.common.aggregate.YnType;
+import sansam.team.common.aggregate.entity.BaseTimeEntity;
 import sansam.team.project.command.application.dto.project.ProjectMemberUpdateDTO;
 
 @Entity
